+++ plugins/woocommerce/tests/e2e-pw/playwright.config.js
@@ -50,14 +50,14 @@ const reporter = [
 		},
 	],
 	[
-		`${ TESTS_ROOT_PATH }/reporters/environment-reporter.js`,
+		`${ TESTS_ROOT_PATH }/reporters/environment-reporter.ts`,
 		{ outputFolder: `${ TESTS_ROOT_PATH }/test-results/allure-results` },
 	],
 ];
 
 if ( process.env.CI ) {
 	reporter.push( [ 'buildkite-test-collector/playwright/reporter' ] );
-	reporter.push( [ `${ TESTS_ROOT_PATH }/reporters/skipped-tests.js` ] );
+	reporter.push( [ `${ TESTS_ROOT_PATH }/reporters/skipped-tests.ts` ] );
 	reporter.push( [
 		'junit',
 		{
@@ -80,18 +80,18 @@ export const setupProjects = [
 	{
 		name: 'install wc',
 		testDir: `${ TESTS_ROOT_PATH }/fixtures`,
-		testMatch: 'install-wc.setup.js',
+		testMatch: 'install-wc.setup.ts',
 	},
 	{
 		name: 'global authentication',
 		testDir: `${ TESTS_ROOT_PATH }/fixtures`,
-		testMatch: 'auth.setup.js',
+		testMatch: 'auth.setup.ts',
 		dependencies: [ 'install wc' ],
 	},
 	{
 		name: 'site setup',
 		testDir: `${ TESTS_ROOT_PATH }/fixtures`,
-		testMatch: `site.setup.js`,
+		testMatch: `site.setup.ts`,
 		dependencies: [ 'global authentication' ],
 	},
 ];
@@ -130,7 +130,11 @@ export default defineConfig( {
 		...setupProjects,
 		{
 			name: 'e2e',
-			testIgnore: '**/api-tests/**',
+			testIgnore: [
+				'**/api-tests/**',
+				/* Exclude PayPal tests, as they don't run well in parallel - see https://github.com/woocommerce/woocommerce/pull/63068. */
+				'**/tests/paypal/**',
+			],
 			dependencies: [ 'site setup' ],
 		},
 		{
@@ -141,6 +145,11 @@ export default defineConfig( {
 		{
 			name: 'legacy-mini-cart',
 			testMatch: [ '**/tests/cart/**', '**/tests/checkout/**' ],
+			dependencies: [ 'site setup' ],
+		},
+		{
+			name: 'paypal-standard',
+			testMatch: [ '**/tests/paypal/**' ],
 			dependencies: [ 'site setup' ],
 		},
 	],